Refactors all of the image processing out from Pdf.java to ImageUtils.java
