+++ src/main/java/edu/ncsu/dlf/servlet/FileUploadServlet.java
@@ -8,18 +8,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.eclipse.egit.github.core.Repository;
-import org.eclipse.egit.github.core.User;
-import org.eclipse.egit.github.core.client.GitHubClient;
-import org.eclipse.egit.github.core.service.RepositoryService;
-import org.eclipse.egit.github.core.service.UserService;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
 import org.apache.commons.codec.binary.Base64;
-import org.apache.commons.io.FileUtils;
-import java.io.File;
 import java.io.InputStream;
 import java.io.ByteArrayInputStream;
 
@@ -28,7 +17,6 @@ import edu.ncsu.dlf.model.PdfComment;
 
 public class FileUploadServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
-
 
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
@@ -39,7 +27,7 @@ public class FileUploadServlet extends HttpServlet {
 		InputStream fileStream = new ByteArrayInputStream(data);
 		//FileUtils.writeByteArrayToFile(new File("test.pdf"), data);
 
-		Pdf test = new Pdf(fileStream);
+		Pdf test = new Pdf(fileStream, getServletContext());
 		List<PdfComment> comments = test.getPDFComments();
 
 		for(PdfComment comment: comments) {
